+++ src/test/java/model/LottoMatcherTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -6,9 +6,6 @@ import org.junit.jupiter.api.Test;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import model.Lotto;
-import model.LottoMatcher;
 
 class LottoMatcherTest {
