+++ src/test/java/model/LottoResultTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package model;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,9 +7,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import model.Prize;
-import model.LottoResult;
 
 class LottoResultTest {
