+++ src/test/java/model/LottoTest.java
@@ -1,6 +1,5 @@
-package lotto;
+package model;
 
-import model.Lotto;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
