+++ src/test/java/util/ParserTest.java
@@ -1,11 +1,9 @@
-package lotto;
+package util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
-
-import util.Parser;
 
 class ParserTest {
